Add Application(TP and OB)

+++ application/src/main/java/common/datatype/internal/AvgVehicleSpeedTuple.java
@@ -170,8 +170,8 @@ public final class AvgVehicleSpeedTuple extends common.util.datatypes.StreamValu
      * @return the VID of this tuple
      */
     @Override
-    public final Short getSegment() {
-        return (Short) super.get(SEG_IDX);
+    public final int getSegment() {
+        return (Integer) super.get(SEG_IDX);
     }
 
     /**
